move enterprise/internal/paths to internal/paths
