+++ enterprise/internal/own/codeowners/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//enterprise/internal/own/codeowners/v1:codeowners",
-        "//enterprise/internal/paths",
+        "//internal/paths",
         "//internal/api",
         "//internal/lazyregexp",
         "//internal/types",
