+++ enterprise/internal/own/codeowners/file.go
@@ -4,8 +4,8 @@ import (
 	"sync"
 
 	codeownerspb "github.com/sourcegraph/sourcegraph/enterprise/internal/own/codeowners/v1"
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/paths"
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/paths"
 )
 
 type RulesetSource interface {
