+++ internal/embeddings/embed/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//internal/embeddings/embed/client",
         "//internal/embeddings/embed/client/openai",
         "//internal/embeddings/embed/client/sourcegraph",
-        "//enterprise/internal/paths",
+        "//internal/paths",
         "//internal/api",
         "//internal/binary",
         "//internal/codeintel/types",
