+++ internal/paths/BUILD.bazel
@@ -4,7 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "paths",
     srcs = ["match.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/paths",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/paths",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
         "//lib/errors",
